Fix mypy errors

+++ colorcode/gradient_curve.py
@@ -36,7 +36,7 @@ class ExponentialCurve(GradientCurve):
         self._exponent = exponent
 
     def get_curve_value(self, x: float) -> float:
-        return x**self._exponent
+        return float(x**self._exponent)
 
 
 ###############################################################################
@@ -61,4 +61,4 @@ class AgnesiWitchCurve(GradientCurve):
         self._factor = factor
 
     def get_curve_value(self, x: float) -> float:
-        return (8 * (self._factor**2.2154)) / ((x - 0.5) ** 2 + (4 * self._factor**2))
+        return float((8 * (self._factor**2.2154)) / ((x - 0.5) ** 2 + (4 * self._factor**2)))
